Update schedule name matches
* match to influenza immunization when product name contains "influenza
  vaccine"
* match to MMR immunization when product name contains mmrv
* match to HIB immunization when product name contains 'haemophilus influenza type b'

Going fwd, I'd like to prioritize the ability to pass in custom
schedules when the JS lib is initialized. That'll make it much more
useful for folks that want to use it for their own needs w/ unique
naming and coding conventions. Also that functionality could support
creating per patient custom schedules...